add option to use double or single precision waves

+++ VC8/utils.h
@@ -65,6 +65,6 @@ void mySetWaveNote(std::string waveNote, waveHndl waveHandle);
 std::string getStandardWaveNote(int brickletID);
 
 void waveClearNaN64(double *wavePtr, long length);
-void waveClearNaN32(double *wavePtr, long length);
+void waveClearNaN32(float  *wavePtr, long length);
 
 #endif // UTILS_H
